+++ extension.js
@@ -11,21 +11,24 @@
 
 'use strict';
 
-const {Clutter, Graphene, GObject, Shell, St, Meta, Gio} = imports.gi;
+import Gio from 'gi://Gio';
+import Meta from 'gi://Meta';
+import Clutter from 'gi://Clutter';
+import Graphene from 'gi://Graphene';
+import GObject from 'gi://GObject';
+import Shell from 'gi://Shell';
+import St from 'gi://St';
 
-const Util           = imports.misc.util;
-const Main           = imports.ui.main;
-const Layout         = imports.ui.layout;
-const WorkspacesView = imports.ui.workspacesView.WorkspacesView;
-const FitMode        = imports.ui.workspacesView.FitMode;
-const WorkspaceAnimationController =
-  imports.ui.workspaceAnimation.WorkspaceAnimationController;
+import * as Util from 'resource:///org/gnome/shell/misc/util.js';
+import * as Main from 'resource:///org/gnome/shell/ui/main.js';
+import {PressureBarrier} from 'resource:///org/gnome/shell/ui/layout.js';
+import {WorkspacesView, FitMode} from 'resource:///org/gnome/shell/ui/workspacesView.js';
+import {WorkspaceAnimationController} from 'resource:///org/gnome/shell/ui/workspaceAnimation.js';
+import {Extension} from 'resource:///org/gnome/shell/extensions/extension.js';
 
-const ExtensionUtils = imports.misc.extensionUtils;
-const Me             = imports.misc.extensionUtils.getCurrentExtension();
-const utils          = Me.imports.src.utils;
-const DragGesture    = Me.imports.src.DragGesture.DragGesture;
-const Skybox         = Me.imports.src.Skybox.Skybox;
+import * as utils from './src/utils.js';
+import {DragGesture} from './src/DragGesture.js';
+import {Skybox} from './src/Skybox.js';
 
 //////////////////////////////////////////////////////////////////////////////////////////
 // This extensions tweaks the positioning of workspaces in overview mode and while      //
@@ -38,11 +41,8 @@ const MAX_VERTICAL_ROTATION = 50;
 // Spacing to the screen sides of the vertically rotated cube.
 const PADDING_V_ROTATION = 0.2;
 
-class Extension {
-  // The constructor is called once when the extension is loaded, not enabled.
-  constructor() {
-    this._lastWorkspaceWidth = 0;
-  }
+export default class DesktopCube extends Extension {
+  _lastWorkspaceWidth = 0;
 
   // ------------------------------------------------------------------------ public stuff
 
@@ -51,7 +51,7 @@ class Extension {
   enable() {
 
     // Store a reference to the settings object.
-    this._settings = ExtensionUtils.getSettings();
+    this._settings = this.getSettings();
 
     // We will monkey-patch these methods. Let's store the original ones.
     this._origUpdateWorkspacesState = WorkspacesView.prototype._updateWorkspacesState;
@@ -61,39 +61,8 @@ class Extension {
     this._origPrepSwitch = WorkspaceAnimationController.prototype._prepareWorkspaceSwitch;
     this._origFinalSwitch = WorkspaceAnimationController.prototype._finishWorkspaceSwitch;
 
-    // We may also override these animation times.
-    this._origWorkspaceSwitchTime = imports.ui.workspacesView.WORKSPACE_SWITCH_TIME;
-    this._origToOverviewTime      = imports.ui.overview.ANIMATION_TIME;
-    this._origToAppDrawerTime = imports.ui.overviewControls.SIDE_CONTROLS_ANIMATION_TIME;
-
-    // We will use extensionThis to refer to the extension inside the patched methods of
-    // the WorkspacesView.
+    // We will use extensionThis to refer to the extension inside the patched methods.
     const extensionThis = this;
-
-    // Connect the animation times to our settings.
-    const loadAnimationTimes = () => {
-      {
-        const t = this._settings.get_int('overview-transition-time');
-        imports.ui.overview.ANIMATION_TIME = (t > 0 ? t : this._origToOverviewTime);
-      }
-      {
-        const t = this._settings.get_int('appgrid-transition-time');
-        imports.ui.overviewControls.SIDE_CONTROLS_ANIMATION_TIME =
-          (t > 0 ? t : this._origToAppDrawerTime);
-      }
-      {
-        const t = this._settings.get_int('workspace-transition-time');
-        imports.ui.workspacesView.WORKSPACE_SWITCH_TIME =
-          (t > 0 ? t : this._origWorkspaceSwitchTime);
-      }
-    };
-
-    this._settings.connect('changed::overview-transition-time', loadAnimationTimes);
-    this._settings.connect('changed::appgrid-transition-time', loadAnimationTimes);
-    this._settings.connect('changed::workspace-transition-time', loadAnimationTimes);
-
-    loadAnimationTimes();
-
 
     // -----------------------------------------------------------------------------------
     // ------------------------------- cubify the overview -------------------------------
@@ -598,8 +567,7 @@ class Extension {
     // ignore this parameter. Instead, we check for the correct action mode in the trigger
     // handler.
     this._pressureBarrier =
-      new Layout.PressureBarrier(this._settings.get_int('edge-switch-pressure'),
-                                 Layout.HOT_CORNER_PRESSURE_TIMEOUT, 0);
+      new PressureBarrier(this._settings.get_int('edge-switch-pressure'), 1000, 0);
 
     // Update pressure threshold when the corresponding settings key changes.
     this._settings.connect('changed::edge-switch-pressure', () => {
@@ -763,10 +731,6 @@ class Extension {
     WorkspaceAnimationController.prototype.animateSwitch = this._origAnimateSwitch;
     WorkspaceAnimationController.prototype._prepareWorkspaceSwitch = this._origPrepSwitch;
     WorkspaceAnimationController.prototype._finishWorkspaceSwitch = this._origFinalSwitch;
-
-    imports.ui.workspacesView.WORKSPACE_SWITCH_TIME = this._origWorkspaceSwitchTime;
-    imports.ui.overview.ANIMATION_TIME              = this._origToOverviewTime;
-    imports.ui.overviewControls.SIDE_CONTROLS_ANIMATION_TIME = this._origToAppDrawerTime;
 
     // Remove all drag-to-rotate gestures.
     this._removeDesktopDragGesture();