+++ src/Skybox.js
@@ -11,13 +11,15 @@
 
 'use strict';
 
-const {Clutter, Gio, GObject, GdkPixbuf, Cogl, Shell} = imports.gi;
+import Gio from 'gi://Gio';
+import Clutter from 'gi://Clutter';
+import GObject from 'gi://GObject';
+import GdkPixbuf from 'gi://GdkPixbuf';
+import Cogl from 'gi://Cogl';
 
-const Main = imports.ui.main;
+import * as Main from 'resource:///org/gnome/shell/ui/main.js';
 
-const ExtensionUtils = imports.misc.extensionUtils;
-const Me             = imports.misc.extensionUtils.getCurrentExtension();
-const utils          = Me.imports.src.utils;
+import * as utils from './utils.js';
 
 //////////////////////////////////////////////////////////////////////////////////////////
 // This file contains two classes, the Skybox (which is an actor) and the SkyboxEffect, //
@@ -37,7 +39,7 @@ var SkyboxEffect = GObject.registerClass({
       'pitch': GObject.ParamSpec.double('pitch', 'pitch', 'pitch', GObject.ParamFlags.READWRITE,
                                         -0.5 * Math.PI, 0.5 * Math.PI, 0),
     },
-  }, class SkyboxEffect extends Shell.GLSLEffect {
+}, class SkyboxEffect extends Clutter.Effect {
   // clang-format on
   _init(file) {
     super._init();
@@ -57,14 +59,16 @@ var SkyboxEffect = GObject.registerClass({
     // Redraw if either the pitch or the yaw changes.
     this.connect('notify::yaw', () => {this.queue_repaint()});
     this.connect('notify::pitch', () => {this.queue_repaint()});
-  };
 
-  // This is called once to setup the Cogl.Pipeline.
-  vfunc_build_pipeline() {
+    const backend     = Clutter.get_default_backend();
+    const coglContext = backend.get_cogl_context();
+
+    this._pipeline = new Cogl.Pipeline(coglContext);
 
     // In the vertex shader, we compute the view space position of the actor's corners.
-    this.add_glsl_snippet(Shell.SnippetHook.VERTEX, 'varying vec4 vsPos;',
-                          'vsPos = cogl_modelview_matrix * cogl_position_in;', false);
+    this._pipeline.add_snippet(
+      Cogl.Snippet.new(Cogl.SnippetHook.VERTEX, 'varying vec4 vsPos;',
+                       'vsPos = cogl_modelview_matrix * cogl_position_in;'));
 
     const fragmentDeclares = `
       varying vec4      vsPos;
@@ -99,19 +103,28 @@ var SkyboxEffect = GObject.registerClass({
       cogl_color_out = texture2D(uTexture, vec2(x, y));
     `;
 
-    this.add_glsl_snippet(Shell.SnippetHook.FRAGMENT, fragmentDeclares, fragmentCode,
-                          false);
-  }
+    this._pipeline.add_snippet(
+      Cogl.Snippet.new(Cogl.SnippetHook.FRAGMENT, fragmentDeclares, fragmentCode));
+  };
 
   // For each draw call, we have to set some uniform values.
-  vfunc_paint_target(node, paintContext) {
+  vfunc_paint_node(node, paintContext) {
     if (this._texture) {
-      this.get_pipeline().set_layer_texture(0, this._texture.get_texture());
-      this.set_uniform_float(this.get_uniform_location('uTexture'), 1, [0]);
-      this.set_uniform_float(this.get_uniform_location('uPitch'), 1, [this.pitch]);
-      this.set_uniform_float(this.get_uniform_location('uYaw'), 1, [this.yaw]);
+      this._pipeline.set_layer_texture(0, this._texture.get_texture());
 
-      super.vfunc_paint_target(node, paintContext);
+      this._pipeline.set_uniform_1i(this._pipeline.get_uniform_location('uTexture'), 0);
+      this._pipeline.set_uniform_1f(this._pipeline.get_uniform_location('uPitch'),
+                                    this.pitch);
+      this._pipeline.set_uniform_1f(this._pipeline.get_uniform_location('uYaw'),
+                                    this.yaw);
+
+      const pipelineNode = new Clutter.PipelineNode(this._pipeline);
+      pipelineNode.set_name('Dekstop-Cube Skybox');
+      node.add_child(pipelineNode);
+
+      const actor = this.get_actor();
+      const alloc = actor.get_allocation_box();
+      pipelineNode.add_texture_rectangle(alloc, 0, 0, 1, 1);
     }
   }
 
@@ -148,7 +161,7 @@ var SkyboxEffect = GObject.registerClass({
 //////////////////////////////////////////////////////////////////////////////////////////
 
 // clang-format off
-var Skybox = GObject.registerClass({
+export var Skybox = GObject.registerClass({
     Properties: {
       'yaw':   GObject.ParamSpec.double('yaw', 'yaw', 'yaw', GObject.ParamFlags.READWRITE,
                                         -2 * Math.PI, 2 * Math.PI, 0),