+++ src/utils.js
@@ -11,9 +11,6 @@
 
 'use strict';
 
-const ExtensionUtils = imports.misc.extensionUtils;
-const Me             = imports.misc.extensionUtils.getCurrentExtension();
-
 // This method can be used to write a message to GNOME Shell's log. This is enhances
 // the standard log() functionality by prepending the extension's name and the location
 // where the message was logged. As the extensions name is part of the location, you
@@ -21,15 +18,25 @@ const Me             = imports.misc.extensionUtils.getCurrentExtension();
 // journalctl -f -o cat | grep -E 'desktop-cube|'
 // This method is based on a similar script from the Fly-Pie GNOME Shell extension which
 // os published under the MIT License (https://github.com/Schneegans/Fly-Pie).
-function debug(message) {
+export function debug(message) {
   const stack = new Error().stack.split('\n');
 
   // Remove debug() function call from stack.
   stack.shift();
 
-  // Find the index of the extension directory (e.g. desktopcube@schneegans.github.com)
+  // Find the index of the extension directory (e.g. desktop-cube@schneegans.github.com)
   // in the stack entry. We do not want to print the entire absolute file path.
-  const extensionRoot = stack[0].indexOf(Me.metadata.uuid);
+  const extensionRoot = stack[0].indexOf('desktop-cube@schneegans.github.com');
 
-  log('[' + stack[0].slice(extensionRoot) + '] ' + message);
+  console.log('[' + stack[0].slice(extensionRoot) + '] ' + message);
+}
+
+// This method can be used to import gettext. This is done differently in the
+// GNOME Shell process and in the preferences process.
+export async function importGettext() {
+  if (typeof global === 'undefined') {
+    return (await import('resource:///org/gnome/Shell/Extensions/js/extensions/prefs.js'))
+      .gettext;
+  }
+  return (await import('resource:///org/gnome/shell/extensions/extension.js')).gettext;
 }